merge: 구조 병합
