--- FE/src/actions/filterAction.js
@@ -1,28 +1,0 @@
-export const configureDate = data => {
-    return {
-        type: "CONFIGURE_DATE",
-        data
-    };
-};
-
-export const configureGuest = data => {
-    return {
-        type: "CONFIGURE_GUEST",
-        data
-    };
-};
-
-export const configurePrice = data => {
-    return {
-        type: "CONFIGURE_PRICE",
-        data
-    };
-};
-
-export const configurePriceDistribution = data => {
-    return {
-        type: "CONFIGURE_PRICE_DISTRIBUTION",
-        data
-    };
-};
-
