+++ FE/src/index.jsx
@@ -4,11 +4,10 @@ import App from '@/views/App';
 
 import { createStore } from 'redux';
 import { Provider } from 'react-redux';
-import { filterReducer } from './reducers/filterReducer';
 import rootReducer from './modules';
 
 const store = createStore(rootReducer);
-// const store = createStore(filterReducer);
+
 window.store = store;
 
 ReactDOM.render(
